backend/requirements.txt
inference-sdk
fastapi
uvicorn[standard]
python-multipart
pillow
openai
python-dotenv
opencv-python
mediapipe
numpy

